Support for 128Mbit SPI flash

+++ src/main/drivers/flash.h
@@ -20,7 +20,7 @@
 #include <stdint.h>
 
 typedef struct flashGeometry_t {
-    uint8_t sectors; // Count of the number of erasable blocks on the device
+    uint16_t sectors; // Count of the number of erasable blocks on the device
 
     uint16_t pagesPerSector;
     uint16_t pageSize; // In bytes
